adjusted for minimizer triplets hashed to integers, using two phases

+++ src/test/java/org/gobiws26/utils/KmerIteratorTests.java
@@ -21,7 +21,7 @@ public class KmerIteratorTests {
 
     @Test
     public void MinimizerWorks() {
-        short minimizer = Minimizers.getMinimizer(firstKmer); // expected: ACGCATGC -> 0001110100101101
-        assertEquals((short) Integer.parseInt("0001110100101101", 2), minimizer);
+        int minimizer = Minimizers.getMinimizer(firstKmer); // expected: ACGCATGC -> 0001110100101101
+        assertEquals(Integer.parseInt("0001110100101101", 2), minimizer);
     }
 }
